Separated Header component from album and pictures pages.
